+++ src/router/src/apis/authentication/mod.rs
@@ -1,3 +1,3 @@
-pub mod api;
+pub mod authentication_api;
 pub mod data;
 pub mod util;
