+++ src/router/src/apis/user/mod.rs
@@ -1,1 +1,2 @@
-pub mod api;
+pub mod user_api;
+pub mod user_data;
