+++ src/router/src/data/mod.rs
@@ -1,4 +1,3 @@
-pub mod auth_data;
 pub mod code;
 pub mod response;
 pub mod user;
